requirements.txt
# Core Dependencies
Flask==3.0.0
python-dotenv==1.0.0
PyYAML==6.0.1

# RabbitMQ
pika==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# SSH (for remote GPSS operations)
paramiko==3.4.0

# HTTP Requests
requests==2.31.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Web & WebSocket
Flask-CORS==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Utilities
colorama==0.4.6

# Logging
loguru==0.7.2

# ML
onnx==1.14.0
skl2onnx==1.15.0
onnxruntime==1.16.0
joblib==1.3.2  # For loading models saved by kdd_intrusion_detection.py
